multiple small changes
remove patch the patch that affected gnat_switches.py as it pulled in the
entire gcc mirror and we don't really need it and we should do it better
when/if we do need it.

rename install-prerequisistes as it's probably better to be consistent with
other <blah>-install.

force the make (all) and make install (all-install) to do everything in
order for no good reason except "just in case".

don't have make release-install do an automatic download, instead do seperate
make release-download.  This is to preserve the ability to do "sudo make release-install without too much confusion with permissions in the release folder.