add gnat-prefix because make prefix=<install-dir> corrupts spark2014/why3 build
make command line definitions are preserved in MAKEFLAGS. spark2014 makefile
wants to install why3 in the build tree not in prefix.  why3 gets confused.
Work around is to remove prefix from the environment (previous fix) and not
indirectly add a prefix=<install-dir> declaration to MAKEFLAGS either.

TODO: update documentation to include gnat-prefix and deprecate or replace
prefix with gnat-prefix (or something else) in both documentation and Makefile.