new release - move to Ubuntu Hirsute
The move to Hirsute was prompted by changes in GTK requiring
features in pango not supported in Ubuntu Focal.  The move
itself required moving to python3 for cairo and pyobject.
As a result all references and dependencies for python2 were
changed to python3 in the build.  Also Hirsute upgraded
it's default llvm version and so that is changed as well.

At this time gps doesn't build because of a compiler bug and
spark2014 doesn't build because of a missing adacore symbol.
There are likely other problems as both the gps and spark2014
builds have been broken since March.